requirements.txt
ultralytics
streamlit==1.29.0
opencv-python
pillow
torch
torchvision
